added navigation
made some navigation thingies
